start typing engine, and clear up some type names
