+++ src/App.tsx
@@ -1,13 +1,13 @@
 import "./styles.css";
 import pako from "pako";
 import { useState, useEffect } from "react";
-import { run, getTooltips, TICK_RATE } from "./Engine.ts";
+import { run, getTooltips, TICK_RATE, GameState, BoardCard } from "./Engine.ts";
 
 // Import JSON files
 import cardsData from "./json/v2_Cards.json";
 import encountersData from "./json/encounterDays.json";
 // Import JSON types generated by quicktype
-import { V2Cards } from "./types/cardTypes.ts";
+import { Tier, V2Card, V2Cards } from "./types/cardTypes.ts";
 import { EncounterDays } from "./types/encounterTypes.ts";
 
 /* Bugs
@@ -61,50 +61,6 @@ console.log(Encounters);
 
 const CardsValues = Object.values(Cards);
 
-interface GameState {
-  tick: number;
-  isPlaying: boolean;
-  players: Player[];
-  multicast: any[];
-  getRand: () => number;
-}
-
-interface Player {
-  HealthMax: number;
-  Health: number;
-  HealthRegen: number;
-  Shield: number;
-  Burn: number;
-  Poison: number;
-  Gold: number;
-  Income: number;
-  board: (BoardCard | BoardSkill)[];
-}
-
-interface BoardCard {
-  card: any;
-  tick: number;
-  Slow: number;
-  Freeze: number;
-  Haste: number;
-  CritChance: number;
-  DamageCrit: number;
-  tier: string;
-  Ammo?: number;
-  DamageAmount?: number;
-  HealAmount?: number;
-  BurnApplyAmount?: number;
-  PoisonApplyAmount?: number;
-  ShieldApplyAmount?: number;
-  [key: string]: any;
-}
-
-interface BoardSkill {
-  card: any;
-  tier: string;
-  [key: string]: any;
-}
-
 const initialGameState: GameState = {
   tick: 0,
   isPlaying: true,
@@ -135,15 +91,11 @@ const initialGameState: GameState = {
   getRand: sfc32(0, 10000, 10000000, 100000000000)
 };
 
-/**
- * Creates a board card from a card object and a specified tier.
- * If the given tier does not exist in the card's Tiers, the first available tier is used.
- *
- * @param {Object} card - The card object from CardsValues.
- * @param {string} tier - The desired tier.
- * @returns {Object} The board card object.
- */
-function _createBoardCardFromCard(card, tier, enchantment) {
+function _createBoardCardFromCard(
+  card: V2Card,
+  tier: Tier,
+  enchantment
+): BoardCard {
   let attributes = {
     Abilities: card.Abilities,
     Auras: card.Auras,
